Added OrderController methods

+++ src/main/java/com/example/demo/jpa/CustomerRepos.java
@@ -3,5 +3,5 @@ package com.example.demo.jpa;
 import com.example.demo.models.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepos extends JpaRepository<CustomerEntity,Integer> {
+public interface CustomerRepos extends JpaRepository<CustomerEntity,Long> {
 }
